requirements.txt
flask==1.1.2
pymongo==3.12.0
scikit-learn==0.24.2
pandas==1.2.5
numpy==1.20.2
rdkit-pypi==2021.9.4